day10/password_security_system/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.4
aioredis==2.0.1
email-validator==2.1.1
zxcvbn==4.4.28
slowapi==0.1.9
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
fakeredis==2.21.3
